# resource/src/swagger/resource-api.yaml
openapi: 3.0.1
info:
    title: course-service
    description: This is a sample course microservice
    version: 1.0.0
servers:
    - url: localhost
paths:
    /api/resource/{id}:
        get:
            tags:
                - resource
            summary: get a single resource by id
            operationId: getResource
            parameters:
                - $ref: '#/components/parameters/id'
            responses:
                200:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Resource'
                400:
                    description: unsuccesful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    errors:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Errors'

    /api/resource/all:
        get:
            tags:
                - resource
            summary: retrieve all resource from database
            operationId: getAllResource
            responses:
                200:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Resource'

                400:
                    description: unsuccesful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    errors:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Errors'
    /api/resource/destroy:
        post:
            tags:
                - resource
            summary: delete a resource by id
            operationId: deleteResource
            requestBody:
                description: delete resource from the backend
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                id:
                                    type: string
            responses:
                202:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        type: boolean

                400:
                    description: unsuccesful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    errors:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Errors'
    /api/resource/add:
        post:
            tags:
                - resource
            summary: add a resource
            operationId: addResource
            requestBody:
                description: add resource to the backend
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                name:
                                    type: string
                                type:
                                    type: string
                                learningStatus:
                                    type: number
                                description:
                                    type: string
                                    nullable: true
                                skillId:
                                    type: array
                                    nullable: true
                                    items:
                                        $ref: '#/components/schemas/SkillData'

            responses:
                201:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Resource'

                400:
                    description: unsuccesful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    errors:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Errors'
    /api/resource/update:
        post:
            tags:
                - resource
            summary: edit a resource
            operationId: EditResource
            requestBody:
                description: edit resource in the backend
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                id:
                                    type: string
                                name:
                                    type: string
                                type:
                                    type: string
                                learningStatus:
                                    type: number
                                description:
                                    type: string
                                    nullable: true
                                skillId:
                                    type: array
                                    nullable: true
                                    items:
                                        $ref: '#/components/schemas/SkillData'

            responses:
                201:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Resource'

                400:
                    description: unsuccesful operation
                    content:
                        application/json:
                            schema:
                                properties:
                                    errors:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Errors'
components:
    parameters:
        id:
            name: id
            in: path
            description: course Id values that need to be considered for filter
            required: true
            schema:
                type: string
    schemas:
        Errors:
            type: object
            properties:
                message:
                    type: string
                field:
                    nullable: true
                    type: string
        Resource:
            type: object
            properties:
                _id:
                    type: string
                userId:
                    type: string
                name:
                    type: string
                type:
                    type: string
                    nullable: true
                learningStatus:
                    type: number
                version:
                    type: number
                description:
                    type: string
                skillId:
                    type: array
                    items:
                        type: string
                dbStatus:
                    type: string

        SkillData:
            type: object
            nullable: true
            properties:
                _id:
                    type: string